Applying comments
